Add separate settings files for develop and testing env
