+++ .env.develop.example
@@ -2,6 +2,7 @@
 # Subset of configuration options, with ready defaults for local development
 # of the application. See ".env.example" for full details on each of these options
 
+DJANGO_SETTINGS_MODULE=contactdb.settings.develop
 DEBUG=True
 
 SECRET_KEY=ZoteTheMighty
