+++ .env.example
@@ -1,3 +1,5 @@
+DJANGO_SETTINGS_MODULE=contactdb.settings.settings
+
 # ==== Tag used for images (optional, default master)
 # Only used in docker compose interpolation, not in the app iteself.
 # See also available versions here:
