+++ contactdb/settings/settings.py
@@ -16,7 +16,7 @@ from pathlib import Path
 import environ
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
-BASE_DIR = Path(__file__).resolve().parent.parent
+BASE_DIR = Path(__file__).resolve().parent.parent.parent
 
 env = environ.Env()
 
@@ -295,37 +295,3 @@ if SENTRY_DSN:
             DjangoIntegration(),
         ],
     )
-
-if DEBUG:
-    DJANGO_DEBUG_TOOLBAR = env.bool("DJANGO_DEBUG_TOOLBAR", default=True)
-
-    INTERNAL_IPS = ["127.0.0.1"]
-
-    if DJANGO_DEBUG_TOOLBAR:
-        INSTALLED_APPS += ["debug_toolbar"]
-
-    if DEBUG and DJANGO_DEBUG_TOOLBAR:
-        import socket
-
-        def show_toolbar(request):
-            return DEBUG and (
-                request.META.get("REMOTE_ADDR") in INTERNAL_IPS
-                or request.META.get("HTTP_X_REAL_IP") in INTERNAL_IPS
-            )
-
-        hostname, _, ips = socket.gethostbyname_ex(socket.gethostname())
-        INTERNAL_IPS += [".".join(ip.split(".")[:-1] + ["1"]) for ip in ips]
-        MIDDLEWARE.insert(2, "debug_toolbar.middleware.DebugToolbarMiddleware")
-        DEBUG_TOOLBAR_CONFIG = {
-            "SHOW_COLLAPSED": True,
-            "SHOW_TOOLBAR_CALLBACK": f"{__name__}.show_toolbar",
-        }
-
-    AUTH_PASSWORD_VALIDATORS = [
-        {
-            "NAME": "django.contrib.auth.password_validation.MinimumLengthValidator",
-            "OPTIONS": {
-                "min_length": 4,
-            },
-        },
-    ]
