+++ contactdb/settings/testing.py
@@ -1,0 +1,3 @@
+from .settings import *
+
+TESTING = True
