+++ pytest.ini
@@ -1,5 +1,7 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = contactdb.settings
 filterwarnings =
     ignore::UserWarning
     ignore::DeprecationWarning
+
+addopts:
+    --ds=contactdb.settings.testing
